Fix for https://github.com/wso2/product-iots/issues/1504
