+++ client/client/src/main/java/org/wso2/iot/agent/services/operation/OperationManagerBYOD.java
@@ -29,7 +29,6 @@ import android.os.Build;
 import android.support.annotation.RequiresApi;
 import android.support.v4.app.NotificationCompat;
 import android.util.Log;
-import android.widget.Toast;
 
 import org.json.JSONArray;
 import org.json.JSONException;
@@ -37,8 +36,6 @@ import org.json.JSONObject;
 import org.wso2.iot.agent.AlertActivity;
 import org.wso2.iot.agent.AndroidAgentException;
 import org.wso2.iot.agent.R;
-import org.wso2.iot.agent.activities.ServerConfigsActivity;
-import org.wso2.iot.agent.api.WiFiConfig;
 import org.wso2.iot.agent.beans.AppRestriction;
 import org.wso2.iot.agent.beans.ComplianceFeature;
 import org.wso2.iot.agent.beans.DeviceAppInfo;
@@ -55,7 +52,6 @@ import org.wso2.iot.agent.utils.Preference;
 
 import java.io.File;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Calendar;
 import java.util.List;
 import java.util.Map;
@@ -127,63 +123,6 @@ public class OperationManagerBYOD extends OperationManager {
                     operation_value_error));
             operation.setOperationResponse("Error in operation payload");
             setResponse(operation);
-        }
-    }
-
-    @Override
-    public void wipeDevice(Operation operation) throws AndroidAgentException {
-        String inputPin = null;
-        String savedPin = Preference.getString(getContext(), getContextResources().getString(R.string.shared_pref_pin));
-        JSONObject result = new JSONObject();
-        String ownershipType = Preference.getString(getContext(), Constants.DEVICE_TYPE);
-        if (Constants.DEFAULT_OWNERSHIP != null) {
-            ownershipType = Constants.DEFAULT_OWNERSHIP;
-        }
-        try {
-            JSONObject wipeKey;
-            String status;
-            if (operation.getPayLoad() != null) {
-                wipeKey = new JSONObject(operation.getPayLoad().toString());
-                if (!wipeKey.isNull(getContextResources().getString(R.string.shared_pref_pin))) {
-                    inputPin = (String) wipeKey.get(getContextResources().getString(R.string.shared_pref_pin));
-                }
-            }
-
-            if (Constants.OWNERSHIP_COPE.equals(ownershipType.trim())) {
-                status = getContextResources().getString(R.string.shared_pref_default_status);
-                result.put(getContextResources().getString(R.string.operation_status), status);
-            } else if (Constants.OWNERSHIP_BYOD.equals(ownershipType.trim()) ||
-                    (inputPin != null && savedPin != null && inputPin.trim().equals(savedPin.trim()))) {
-                status = getContextResources().getString(R.string.shared_pref_default_status);
-                result.put(getContextResources().getString(R.string.operation_status), status);
-            } else {
-                status = getContextResources().getString(R.string.shared_pref_false_status);
-                result.put(getContextResources().getString(R.string.operation_status), status);
-            }
-
-            operation.setPayLoad(result.toString());
-
-            if (status.equals(getContextResources().getString(R.string.shared_pref_default_status))) {
-                /*Toast.makeText(getContext(), getContextResources().getString(R.string.toast_message_wipe),
-                        Toast.LENGTH_LONG).show();*/
-                operation.setStatus(getContextResources().getString(R.string.operation_value_completed));
-                getResultBuilder().build(operation);
-
-                if (Constants.DEBUG_MODE_ENABLED) {
-                    Log.d(TAG, "Started to wipe data");
-                }
-            } else {
-                /*Toast.makeText(getContext(), getContextResources().getString(R.string.toast_message_wipe_failed),
-                        Toast.LENGTH_LONG).show();*/
-                operation.setStatus(getContextResources().getString(R.string.operation_value_error));
-                operation.setOperationResponse("Invalid PIN code entered.");
-                getResultBuilder().build(operation);
-            }
-        } catch (JSONException e) {
-            operation.setStatus(getContextResources().getString(R.string.operation_value_error));
-            operation.setOperationResponse("Error in parsing WIPE payload.");
-            getResultBuilder().build(operation);
-            throw new AndroidAgentException("Invalid JSON format.", e);
         }
     }
